+++ src/api/users.ts
@@ -18,7 +18,6 @@ export interface ProfilePaymentOption {
   chain_id?: string,
   price?: number,
   object_id?: string,
-  fep_0837_enabled?: boolean,
 }
 
 interface Source {