+++ CHANGELOG.md
@@ -20,6 +20,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Redirect unauthanticated users from local timeline to landing page if timeline is locked.
 - Convert instance name in welcome text into About page link.
 
+### Removed
+
+- Remove pre-FEP-0837 remote subscription previews.
+
 ### Fixed
 
 - Don't reload notification list after navigating to notifications page.